Add files via upload
Unicorn companies are those which have exceeded a billion dollar market valuation. The dataset for this task contains information of over a thousand unicorn companies with the specifics of their sector, country of operations, year of inception and some of the investors. You will be able to use this information to learn how and when the different firms achieve milestones of this magnitude and provide next step action items for the investment firm.